+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.2] - 2025-07-07
+## [2.0.0] - 2025-07-16
 
 ## Removed
 - `n_lookbacks` (which should be called `n_baselines`) for computing confirmation within the SAS
@@ -18,6 +18,8 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Now retrieves the umd distance to land and use a water mask that excluded 1 km from ocean.
 
 ## Added
+- Ability to use algorithm parameters to supplement runconfig for SDS operations.
+  - It's now a base class to `RunConfigData` and if an external 
 - Updated interface for despeckling to use interpolation to fill nan values within burst area.
   - distmetrics>=1.0.0 - see more details there